requirements.txt
aws-cdk-lib==2.74.0
constructs>=10.0.0
boto3==1.26.104


